+++ src/deviceManager.test.ts
@@ -1,3 +1,4 @@
+// src\deviceManager.test.ts
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -5,6 +6,7 @@ process.argv = ['node', 'matterbridge.test.js', '-logger', 'info', '-matterlogge
 
 import { jest } from '@jest/globals';
 import { AnsiLogger, BLUE, db, er, LogLevel, nf, nt, pl, UNDERLINE, UNDERLINEOFF } from 'node-ansi-logger';
+
 import { Matterbridge } from './matterbridge.js';
 import { MatterbridgeEndpoint } from './matterbridgeEndpoint.js';
 import { DeviceManager } from './deviceManager.js';