+++ src/matterbridgeDynamicPlatform.test.ts
@@ -1,11 +1,12 @@
+// src\matterbridgeDynamicPlatform.test.ts
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 process.argv = ['node', 'matterbridge.test.js', '-frontend', '0', '-profile', 'JestDynamic'];
 
 import { jest } from '@jest/globals';
-
 import { AnsiLogger, LogLevel } from 'node-ansi-logger';
+
 import { Matterbridge } from './matterbridge.js';
 import { MatterbridgeDynamicPlatform } from './matterbridgeDynamicPlatform.js';
 import { MatterbridgePlatform } from './matterbridgePlatform.js';