+++ src/utils/copyDirectory.test.ts
@@ -1,10 +1,7 @@
+// src\utils\copyDirectory.test.ts
 /* eslint-disable @typescript-eslint/no-empty-function */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-// copyDirectory.test.ts
-/**
- * @jest-environment node
- */
 import { jest } from '@jest/globals';
 import { AnsiLogger } from 'node-ansi-logger';