+++ src/utils/utils.test.ts
@@ -1,6 +1,12 @@
+// src\utils\utils.test.ts
 /* eslint-disable @typescript-eslint/no-empty-function */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unused-vars */
+import { jest } from '@jest/globals';
+import { promises as fs } from 'node:fs';
+import path from 'node:path';
+import { AnsiLogger } from 'node-ansi-logger';
+
 import {
   deepEqual,
   deepCopy,
@@ -10,19 +16,24 @@ import {
   waiter,
   wait,
   getMacAddress,
-  createZip,
   getNpmPackageVersion,
   copyDirectory,
   resolveHostname,
   getStringArrayParameter,
   getIntArrayParameter,
   parseVersionString,
-} from './export';
-import { hasParameter, getParameter, getIntParameter, isValidIpv4Address, isValidNumber, isValidBoolean, isValidString, isValidObject, isValidArray, isValidNull, isValidUndefined } from './export';
-import { promises as fs } from 'node:fs';
-import { AnsiLogger } from 'node-ansi-logger';
-import path from 'node:path';
-import { jest } from '@jest/globals';
+  hasParameter,
+  getParameter,
+  getIntParameter,
+  isValidIpv4Address,
+  isValidNumber,
+  isValidBoolean,
+  isValidString,
+  isValidObject,
+  isValidArray,
+  isValidNull,
+  isValidUndefined,
+} from './export.js';
 
 describe('Utils test', () => {
   let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;